training/mobilenetv3-small/requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10.0
pandas>=2.0
scikit-learn>=1.3
tqdm>=4.66
